Move message properties to Message classes
